added edit post, delete post, user page features

+++ server/src/controllers/postsController.js
@@ -37,9 +37,13 @@ module.exports = {
     try {
       const { text } = req.body;
 
-      if (text.length === 0) return res.status(400).json({ error: "Post text can not be empty" });
+      if (text.length === 0)
+        return res.status(400).json({ error: "Post text can not be empty" });
 
-      if (text.length > 500) return res.status(400).json({ error: "Post text must be no longer than 500 characters" });
+      if (text.length > 500)
+        return res
+          .status(400)
+          .json({ error: "Post text must be no longer than 500 characters" });
 
       const newPost = await createPostService(req.body);
 
@@ -55,13 +59,23 @@ module.exports = {
       const postId = req.params.id;
       const { text } = req.body;
 
-      if (text.length === 0) return res.status(400).json({ error: "Post text can not be empty" });
+      const userId = req.user.id;
 
-      if (text.length > 500) return res.status(400).json({ error: "Post text must be no longer than 500 characters" });
+      const post = await getPostService(postId);
 
-      const post = await editPostService(postId, text);
+      if (post.user.id != userId) return res.sendStatus(401);
 
-      return res.status(200).json(post);
+      if (text.length === 0)
+        return res.status(400).json({ error: "Post text can not be empty" });
+
+      if (text.length > 500)
+        return res
+          .status(400)
+          .json({ error: "Post text must be no longer than 500 characters" });
+
+      const editedPost = await editPostService(postId, text);
+
+      return res.status(200).json(editedPost);
     } catch (err) {
       console.error(err);
       return res.status(500).json({ error: "Internal server error" });
@@ -71,6 +85,11 @@ module.exports = {
   async deletePost(req, res) {
     try {
       const postId = req.params.id;
+      const userId = req.user.id;
+
+      const post = await getPostService(postId);
+
+      if (post.user.id != userId) return res.sendStatus(401);
 
       await deletePostService(postId);
 
@@ -84,15 +103,18 @@ module.exports = {
   async createComment(req, res) {
     try {
       const postId = req.params.id;
-      console.log(postId);
       const { userId, text } = req.body;
       const post = await getPostService(postId);
 
       if (!post) return res.status(404).json({ error: "Post not found" });
 
-      if (text.length === 0) return res.status(400).json({ error: "Comment can not be empty" });
+      if (text.length === 0)
+        return res.status(400).json({ error: "Comment can not be empty" });
 
-      if (text.length > 100) return res.status(400).json({ error: "Comment can be on longer than 100 characters" });
+      if (text.length > 100)
+        return res
+          .status(400)
+          .json({ error: "Comment can be on longer than 100 characters" });
 
       const comment = createCommentService(postId, text, userId);
       return res.status(200).json(comment);
